requirements.txt
agate==1.9.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
apache-airflow==2.10.3
apache-airflow-providers-common-compat==1.2.1
apache-airflow-providers-common-io==1.4.2
apache-airflow-providers-common-sql==1.19.0
apache-airflow-providers-fab==1.5.0
apache-airflow-providers-ftp==3.11.1
apache-airflow-providers-http==4.13.2
apache-airflow-providers-imap==3.7.0
apache-airflow-providers-smtp==1.8.0
apache-airflow-providers-sqlite==3.9.0
apispec==6.7.1
argcomplete==3.5.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.1
cachelib==0.9.0
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
crashtest==0.4.1
cron-descriptor==1.4.5
croniter==5.0.1
cryptography==43.0.3
daff==1.3.46
dbt-adapters==1.7.0
dbt-common==1.10.0
dbt-core==1.8.8
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.5.1
deepdiff==7.0.1
Deprecated==1.2.15
dill==0.3.9
distlib==0.3.9
Django==5.1.3
django-cors-headers==4.6.0
django-cors-middleware==1.5.0
django-health-check==3.18.3
django-model-utils==5.0.0
django-storages==1.14.4
djangorestframework==3.15.2
dnspython==2.7.0
drf-yasg==1.21.8
dulwich==0.21.7
email_validator==2.2.0
exceptiongroup==1.2.2
fastjsonschema==2.20.0
filelock==3.16.1
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.10.0
google-re2==1.1.20240702
greenlet==3.1.1
grpcio==1.68.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
inflection==0.5.1
installer==0.7.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kafka-python==2.0.2
keyring==24.3.1
kombu==5.4.2
lazy-object-proxy==1.10.0
leather==0.4.0
limits==3.13.0
linkify-it-py==2.0.3
lockfile==0.12.2
Logbook==1.5.3
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mashumaro==3.14
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
minimal-snowplow-tracker==0.0.2
minio==7.2.10
more-itertools==10.5.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
oauthlib==3.2.2
opentelemetry-api==1.28.1
opentelemetry-exporter-otlp==1.15.0
opentelemetry-exporter-otlp-proto-grpc==1.15.0
opentelemetry-exporter-otlp-proto-http==1.15.0
opentelemetry-proto==1.15.0
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
pexpect==4.9.0
pillow==11.0.0
pkginfo==1.11.2
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prison==0.2.1
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.10.0
pyproject_hooks==1.2.0
PySocks==1.7.1
python-daemon==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-nvd3==0.16.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.21.0
rsa==4.9
selenium==4.26.1
setproctitle==1.3.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.1
starlette==0.41.2
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
tomli==2.0.2
tomlkit==0.13.2
trio==0.27.0
trio-websocket==0.11.1
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uc-micro-py==1.0.3
universal_pathlib==0.2.5
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
virtualenv==20.27.1
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==2.2.3
wirerope==0.4.8
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
xattr==1.1.0
yarl==1.17.2
zipp==3.21.0
